asf: debugging, more error checking, and return useful data
this commit includes some tinkering in analyze_stockdata to figure
out how to get the ticker's rank, group, and group leader.
The acquired data is returned as a dict containing values for 'group',
'rank', and 'leader', or None/np.nan if a value is missing/not
applicable. It will return an empty dict if the urls could not be
accessed.
There are still some debugging print statements in the code that will
eventually be removed.