Major debugging of getCurrentValue/getDateForContext
For some reason (typo, python/jupyter notebooks hates me, I'm
stoopid, ...), having the getDateForContext function as a method in the
SecFiling class did not work. Always got an AttributeError on it when it
was called through a SecFiling10Q object. So I moved it inside the
getCurrentValue method. That's the only place where it should be used,
anyways. Oh well.