Also move getEps, getSales, getCurrentValue into the superclass.
Again, there is pretty much no difference in getting EPS, Sales, etc.
from the different types of filings. The only difference that exists is
the amount of days that pass between consecutive filings of the same
type (365 days for 10-Ks, 91 days for 10-Qs). This is handled by passing
the appropriate values into the super.getCurrentValue() function, and
overriding this function in each subclass to determine the correct
values, and call the super-function with those.
Note that this is currently the only reason to have subclasses for the
different types of filings. I think I like to keep it that way for now,
as an easy way to distinguish between 10-Ks and 10-Qs.