Increase EPS/Sales growth accel to go over 5/4 quarters,
respectively. You need at least three data points to calculate an
acceleration, so using three is somewhat meaningless. According to
Canslim, I want the growth acceleration for the "recent quarters", which
is a little up to interpretation, but using 5 for EPS and 4 for Sales
should show if there were net accelerating (or decelerating) tendencies
over the last year-ish.